Refactor PulsarNode to follow MVVM and improve Uri calls
Moved PulsarNode from Models to ViewModels and renamed to PulsarNodeViewModel, separating concerns and adhering more closely to MVVM. Also, replaced 'new Uri()' with 'new()' in PulsarClient for clearer, more modern syntax. The PulsarService class was similarly transitioned to PulsarTreeModel under Models. Changes were accordingly adjusted in all dependent classes. This makes the code easier to understand and manage, and it improves the organization of the project.