Add Docker configuration and refactor Pulsar connections
Added a Dockerfile, .dockerignore, and updated the docker-compose.yml file for configuring the application in a Dockerized environment. Created a separate Pulsar worker service in docker-compose for better scalability and isolation. Also, made some changes to the ApachePulsarConsumerWorker and ApachePulsarProducerWorker classes to use the dockerized Pulsar service rather than a local one for higher accessibility and easier project setup. Additionally, tweaked a few settings in the PulsarWorker.csproj for Docker compatibility.