Update setting persistence and notification features
Added functionality to persist and update settings in real-time. Changed the method of consuming persisted settings in `SettingsModel` by introducing an `onSuccess` callback function to trigger upon successful setting persistence. The style of this function was also implemented for 'TextSettingViewModel' to enable real-time data persistence with a configurable delay. Moreover, a notification feature is also added, which informs the user about the successful saving of settings. Made necessary updates to the `SettingsViewModel` to include notification services and handle the display of notifications.

Various enhancements also include the use of `ItemsControl` in `Settings.axaml` instead of `ItemsRepeater` for better compatibility, and small fixes in initializing components in Avalonia UI. These changes enhance the performance and user experience.