Upgrade to .NET 8 and update dependencies
The entire solution has been migrated from .NET 7 to .NET 8 to utilize the latest features and improvements. Some of the packages had to be updated to their respective versions that support .NET 8. Furthermore, "AutoMapper.Extensions.Microsoft.DependencyInjection" is replaced by "AutoMapper" package as per the latest documentation for compatibility with .NET 8. A global.json file was also added to ensure .NET 8 SDK is used for future builds.