"Add DbContext and initial configurations for GraphQL in HotChocolate Checker"
Implemented the CheckerDbContext and corresponding configurations for GraphQL in the HotChocolate Checker. Added Entity model for User and service extensions for DB connectivity. Also, established the AutoMapper settings and introduced GraphQL types for Queries and Mutations. These changes are necessary for setting up the database and GraphQL framework in context of the HotChocolate checker application.