Remove unused GraphQL Mutation class and generate a new Initial migration
The redundant Mutation class in the GraphQL layer was removed because it was not being used anywhere in the application.
In addition, a significant number of changes were made to the database model which necessitated a new migration.
Therefore, 'Initial' migration has been created for the EF Core to update the database schema accordingly.