Remove SettingsDependentHttpClient, introduce HttpClientFactory
SettingsDependentHttpClient was removed and HttpClientFactory was introduced in its place to provide better separation of concerns and simplicity. The responsibility of updating the base address has been moved to the classes directly using HttpClient i.e. PulsarClient. This enables on-demand creation and disposal of HttpClient as and when the base address is updated. This update enhances lifecycle management and the overall structure of the application.